Create pvez zones code
If you want a PVP zone marker, in /profiles or /sc go to /PVEZ (I use PVEZ test) add this to the list of zones. I find it best to copy the one above last and edit it then leave the last one without the commer so the file doesn't mess up!